+++ lib/lib_audio/ESP8266Audio/src/libflac/cpu.c
@@ -31,7 +31,7 @@
  */
 
 //#ifdef HAVE_CONFIG_H
-#  include <config.h>
+#  include "config.h"
 //#endif
 
 #include "private/cpu.h"