+++ lib/lib_audio/ESP8266Audio/src/libhelix-aac/assembly.h
@@ -223,7 +223,7 @@ Word64 MADD64(Word64 sum64, int x, int y);
 /* toolchain:           ARM ADS or RealView
  * target architecture: ARM v.4 and above (requires 'M' type processor for 32x32->64 multiplier)
  */
-#elif defined (__arm) && defined (__ARMCC_VERSION)
+#elif defined (XXX__arm) && defined (__ARMCC_VERSION)
 
 static __inline int MULSHIFT32(int x, int y)
 {
@@ -336,7 +336,7 @@ static __inline Word64 MADD64(Word64 sum64, int x, int y)
 /* toolchain:           ARM gcc
  * target architecture: ARM v.4 and above (requires 'M' type processor for 32x32->64 multiplier)
  */
-#elif defined(__GNUC__) && defined(__arm__)
+#elif defined(__GNUC__) && defined(XXXX__arm__)
 
 static inline int MULSHIFT32(int x, int y)
 {