+++ lib/libesp32/Berry/generate/be_fixed_be_class_audio_file_source_fs.h
@@ -1,14 +1,13 @@
 #include "be_constobj.h"
 
 static be_define_const_map_slots(be_class_audio_file_source_fs_map) {
-    { be_const_key(init, 2), be_const_func(i2s_file_source_fs_init) },
     { be_const_key(deinit, -1), be_const_func(i2s_file_source_fs_deinit) },
-    { be_const_key(close, -1), be_const_func(i2s_file_source_fs_deinit) },
+    { be_const_key(init, -1), be_const_func(i2s_file_source_fs_init) },
 };
 
 static be_define_const_map(
     be_class_audio_file_source_fs_map,
-    3
+    2
 );
 
 BE_EXPORT_VARIABLE be_define_const_class(